backend/requirements.txt
SQLAlchemy>=2.0,<3
PyJWT==2.8.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.1
python-escpos==3.0
Pillow>=9.0.0
pyusb>=1.2.1
python-socketio==5.11.0
gunicorn==21.2.0
fastapi==0.115.8
uvicorn[standard]==0.34.0
pytest>=7.0.0
pytest-cov>=4.0.0
